refactor: Shared circleIndex util
